Restructure ui folder
