--- src/client/ui/containers/index.ts
@@ -1,12 +1,0 @@
-import { css } from '@emotion/react';
-
-export const headerStyles = css`
-  height: 9rem;
-  width: 110rem;
-  margin: 1.5rem auto;
-  padding: 0 1.5rem;
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: center;
-  justify-content: space-between;
-`;
